Parse like.js with UltraJSON, use snscraper module
Using asl.literal_eval() was not an option in my case, as it would fail to parse my like.js (which is probably fair enough, as I have over ~200K likes). UltraJSON lets the scraping start more or less instantly on my machine, so seems like a good alternative.

Had issues with getting snscraper to run from CLI so switched to the somewhat more python-y but less well-documented approach of using snscraper as a module.